passes: Move common routines to utils.py
